HYD-4050 DEBUG: Halt a test run if Kernel fails to install correctly
Some research by Brian Murrell has showed that the initramfs needed 
during the reboot process wasn't created during kernel installation.

This patch will allow investigation of that problem by causing the test
run to halt indefinitely if the initramfs is not present on the 
installed node so that the state of the nodes can be inspected by a
member of the development team. 

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: Id51331ad277cef827ba8c46fdab39a103e1f5297
Reviewed-on: http://review.whamcloud.com/13513
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>