TEI-1912: Fix race between node reset and stonith in test framework
Currently, for tests where nodes are configured HA and those nodes
are reset, there is a race between the reset starting the node and
the stonith restarting the node, resulting in a test failure when
stonith wins and the reset errors because the domain is already
active. This change makes the reset not error if the domain is
already started.

Change-Id: I8f46bdfb94c6da0f443b7fa9ca57e9a555f5c11a
Reviewed-on: http://review.whamcloud.com/13555
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>