HYD-3677 Create a better way to report configuration errors to users
Package up the return from the agent so that error messages and errors
can be systematically sent back and given the user in place of stack traces.

This routine puts the principle in place and deals with some of the cases.

Before the next release all the agent calls need to be moved to the new
protocol. Because the protocol is package further changes shoudl be able to
be implemented to all calls easily.

Change-Id: Ia8dec524b8b271ee5fe44dfb32d4cf92f969def0
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/15616
Tested-by: Jenkins
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>