HYD-4076 Upgrade from IEEL 1.0.2.0 to 2.2.0.0 fails
Using models such as ManagedHost directly in migrations is
dangerous because the version of the code base will not be
the same as the database version for migrations that are
for multiple versions.

Imagine a 3 version code base migrating from V1 to V3.

During the V1->V2 migration the code base will actually be
at V3 and this means that if there are fields in V3 that are
not in V2 and the model is used directly (as this code did
before this patch) a field not found error will occur.

For example:
django.db.utils.DatabaseError: column
chroma_core_managedhost.install_method does not exist

Occurs because install_method is required by the code base
but is not yet in the database.

This patch removes the use of the model directly and replaces
it with version safe raw sql.

Change-Id: I71fc39db5c57aa8642311919c0498a9ff789eff7
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/13573
Tested-by: Jenkins
Reviewed-by: Gene Campbell <gene.campbell@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>