HYD-4139 yum change raise inefficiency in the packages scan code.
I6f0418676c10af517f1ae3c9cfd61b4e6f0c81cb introduced an in-efficiency
regression to the way in the yum packages are cleaned.

Rather than just clean out the yum repositories to be scanned for updates,
all the repositories are cleaned out. This will make the check for updates
take longer. The operation does not happen that often and so its not the
end of the world but is worth removing this small regression.

Change-Id: I60a29b33dea926154e1cb272f46733c1750f6681
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/13606
Tested-by: Jenkins
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>