HYD-4389 Enable el6.6 testing by making distribution switchable
The test distribution is currently hardcoded make it possible to
use an environment variable to switch the distribution.

The environment variable is called TEST_DISTRIBUTION and defaults
to el6.6. But the value should in future be set by Jenkins.

This should allow nightlies etc to easily run a matrix of builds.

Change-Id: I0d5ddf8f4906de0358baf5a4caa15fd1ebcdbd2c
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/14628
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-on: http://review.whamcloud.com/13794