HYD-3784 Discovery of newly added OST is non-intuitive Steps
During discovery the manager only displays information about discovered
filesystem but not about discovered targets. This means that someone
discovering for example a new OST will just be presented with a
message saying 'Discovered no new filesystems'.

This patch adds additional messages to say is MGT/OST/MDT's are
discovered or not. Providing a list of what was and wasn't found.

Functionally there are no changes to the way in which the code
operates other than the display of messages.

Change-Id: Ic807f6f4dde0ff6ff33f24c981e65ed6397bc3d4
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/13652
Tested-by: Jenkins
Reviewed-by: Gene Campbell <gene.campbell@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>