HYD-3784 Discovery of newly added OST is unreliable with zfs
There is a window with ZFS where a newly added OST may not be
spotted by the zfs device scan. This is because the delta mechanism
looks for changes in the zfs pools to decide if to send a full scan.

However there is a window between the creation of the zpool and the
creation of the dataset on that zpool which if IML scans the host will
cause it to miss the creation of the dataset.

The agent looks at the out of of 'zpool list' and compares with the
last time to look for changes - if they differ it does a full scan -
however 'zpool list' on lists spools! (Kind of obvious). Changing
that to 'zfs list' will list zpools and datasets, as we don't parse
the output the change really is simple.

So to replicate.

create the zpool and wait for 1 minute before running mkfs.lustre
(which creates the dataset).

This change looks for changes in the datasets and the zpools as the
input to the delta algorithm.

Change-Id: I1b1b3ca4b5f7e875e124585c7440a0c0cf677c37
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/13651
Tested-by: Jenkins
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Gene Campbell <gene.campbell@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>