HYD-4164 IML cannot add OSS/MDS if the servers are using VLAN tagged interface.
We have previously been quite choosy about the names we detect for network connections.
This means that vlans for example (eth0.1) do not work.

The reality seems to be that the name will always be terminated with a whitespace in the
cases where we detect them. For this reason it is safe to just accept all characters up
until the whitespace. Every character will be valid be virtue that a network of the name
detect exists.

Change-Id: I4900cb021b77245a12d401629ad049978386b71a
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/13802
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Gene Campbell <gene.campbell@intel.com>