HYD-4146 Failed to create stonith resource
It seems likely that  there is a condition where by neither of
the nodes in an HA pair configure the stonith part of the
configuration (or actually the some other parts such as
quorum-policy, symmetric-cluster or cluster-infrastructure.).

This can happen without any error happening, meaning that is
would seem to be possible that both nodes can execute
configure_corosync without being the DC.

Whatever the cause of both nodes managing to do this, and
there is one clear case where the DC changes between the
two nodes executing the 'am I the DC' test, the code should
be resilient to this failure.

I've pushed a patch that means if the corosync configuration
doesn't happen then the Corosync configuration will stop and
the installation process will stop. This will allow the user
to restart the setup server step and achieve a successful setup.

The mechanism to do this check is to ensure that at the point in the
configuration where one or the cluster nodes (our clusters are always 2 nodes for this purpose) is expected to have configured stonith, all the nodes check that stonith is configured and exit.

The change is possibly slightly more change than was required
because my original approach was to try and enable a retry but
after discussions it seems to make more sense to detect the
unexpected and stop the configuration process.

Change-Id: Ic131ae9f782d6db9cc798b6e3d7c82c448fb5c6b
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/13734
Tested-by: Jenkins
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Gene Campbell <gene.campbell@intel.com>