HYD-4184 Shell commands on agents hang if the output of  the command (stderr or stdout) is greater than 65536 bytes
A well documented feature of the python process.Popen method is that
if used with PIPE it will hang if the output is greater than 2^16.

This caught us out at customer in the middle east because 'cibadmin --query --local'
was returning 66xxx bytes (not sure of the exact number). The outcome
was that failover failed, although I think other odd behaviours would
possibly occur in the future.

This needs to be fixed up properly by using an intermediate file, this is slower
than using memory but we are probably not that sensitive.

Change-Id: I9ea04c5cdbd90c333bf072650ea825004d5a0ae3
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/13788
Tested-by: Jenkins
Reviewed-by: Gene Campbell <gene.campbell@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>