HYD-1215 Separate lnet state from host state (file moves)
ManagedHost.state is currently overloaded, used for both the lnet state (unloaded, down, up)
and the host configuration state (unconfigured, configured, removed).

This patch is a simple movement of code to get things into the right place before changes are
made. Should allow follow on changes to be easier to managed.

Change-Id: I49b17c3f50364277b099ca2a9912c3881769f9c3
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/15613
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>