HYD-4207 Handling the package conflicts python-pytz vs. pytz
python-pytz was added as a package to the IML bundle for 2.2 this
causes a clash with the standard RH 6.6 installation because we rename
the package to be python-pytz.

This patch adds an ed.spec file to reverse the standard renaming
we have on our python pypi rpms and therefore keeps the package
name as pytz as opposed to python-pytz

The appropriate dependencies are changed in the spec file for
chroma manager and clery.

Change-Id: I888185381ec649852be7e5f474fea6fca480d93f
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/13850
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>