HYD-4210 disconnect modal backdrop not removed in background tab
The "webkitTransitionEnd" event does not fire when the page is in a
background tab, causing the disconnect modal to not be removed.

This leaves a transparent modal background layer, which blocks all
clicks.

By removing animations, we remove the need to listen for the transition
ending, since there will no longer be one.

Change-Id: Ie0653238289bebd6d5288cfc7c772cba92726f10
Reviewed-on: http://review.whamcloud.com/13865
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>