HYD-4215 Update ommitted package dependencies
python-{amqp,pytz,billiard} are not actually required by the manager,
but rather are dependencies of manager dependencies.  We should
limit dependencies to direct dependencies and let those dependencies
manage their own dependencies.

{django,python}-celery and kombu were updated in
13fa52dac7f70a0a6db2eab849e309fb0ff32ec2 but the dependency on it
was not updated.

The manager doesn't actually need python-celery directly but
rather django-celery reqires it.

mod_proxy_wstunnel was updated in
c91136f27485c2fac92673c4900e5a258b261c3a but the dependency on it
was not updated

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: I37a8ec4b38efa991f02155bc0640c66ef355fefa
Reviewed-on: http://review.whamcloud.com/13843
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>