HYD-4231 robinhood-2.5.4 bundle missing robinhood-lhsm rpm
This change actually reverts back to the previous spec file
for HSM and makes the minimal changes required to support 2.5.4.

Our spec file usage and the HSM releases of the spec file are
a fairly mixed bag. Tempting to sort this out now but after
discussions with Brian it seems prudent to just make the minimal
updates to where we are now.

This basically amounts to nothing more than addtion of the
requirement for libattr-devel.

To get a real feeling for the effects of this change it is best
to diff this file with the 6a403c82beb3136e81a8bb94c70feb958b5b5a15
version.

Change-Id: I33f7e803a8ba6febaa203206cd8028e53d17e6e7
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/13910
Tested-by: Jenkins
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Tested-by: Kelsey Prantis <kelsey.prantis@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>