HYD-1215 Separate lnet state from host state.
ManagedHost.state is currently overloaded, used for both the lnet state (unloaded, down, up)
and the host configuration state (unconfigured, configured, removed).

This patch is a refactor of the code and should produce only one external change which is that
the state of lnet now needs to be driven through the lnet object rather than the host object.

This will require a bit of rework in the GUI, other than that the end user functionality is
unchanged.

Change-Id: I01f3109a1c7456a6b80aec13af04530506bb32b9
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/15614
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>