HYD-4161 Implement view-server HYD-4174 Move the about page to the new ui HYD-3063 Most of the IML 2.1 urls malfunction when they don't end with a slash
This patch moves view rendering from Python (Django) to node.
By doing this, we gain total control over how the mocks are fetched and
rendered; allowing us to use stub-daddy to mock them out. Additionally,
it will allow for less context switching when working on the client or
the server-side view layer. Finally, it opens opportunities to start
more server-side rendering / pre-processing. This patch contains a full
suite of unit tests, integration tests will follow in a separate patch.
With the conversion we also moved the about page to the new GUI, and
fixed the bug with trailing slashes being mandatory.

Use a low-level stand-alone node process to handle serving the views.
Right now, views are being served for both the old and new GUIs. As the
old pages disappear we can simply remove the routes for each of the old
pages.

- Remove old Python view system.
- Inline content templates for the new GUI. The client is in charge
of routing.
- Move the about screen to the new GUI.
- Update the old router to work without trailing slashes.
- Update the configuration screens to not require a full page refresh.
- Add XMLRPC lib for retrieving supervisord status.

Change-Id: I223c38604dcc02f80b9bc6a25af05ea0eda3c9dd
Signed-off-by: Joe Grund <joe.grund@intel.com>
Reviewed-on: http://review.whamcloud.com/14309
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Tested-by: Will Johnson <william.c.johnson@intel.com>
Tested-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>