HYD-4293 Remove dead LNet code
We have dead code from the version of LNet config in the
old UI that needs to be removed.

Remove the files that are no longer needed and remove
mention of them in the manager module installer.

Removed the dead scripts, and they are no longer mentioned in setup.py.

Change-Id: I3cc4fde9547c3e8e016411f4bc86f2e6266fc6aa
Signed-off-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Reviewed-on: http://review.whamcloud.com/14161
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>