HYD-4548 Create Req module - A module to assist in making requests
Many modules need the ability to make requests and analyze the results. Sometimes, you need to make a request and only capture / mask part of the results. This is a common pattern we have been using so it makes sense to split this logic out into its own module so that other modules can easily make use of its functionality and eliminate this duplication of code between modules.

The realtime service currently contains a request directory. Most of the logic we need to share between modules is in this folder. This patch extracts the request logic from the realtime module into its own module and condenses it such that it is usable throughout our node modules and is no longer coupled to the realtime module. Note that this patch does NOT update the realtime module to use the new req module. That will be in a separate patch once the req module lands.

  - updated run_tests to include the req module test suite
  - updated cluster_setup to run ziplock for the req module
  - updated the test_dependencies to include the req module
  - Added add-request-info.js, which will push any errors during a request
  - Added buffer-request.js, which allows you to specify the path and options of a request. The result is checked for errors, converted to json, and ran through a json mask.
  - Added build-options.js, which allows you to specify a path and options. The resulting output will be an built object that can be used to make the request.
  - Added error-buffer.js which handles encountered errors
  - Added the main index.js file, which handles making the request
  - Added jasmine configuration for testing
  - Added mask.js to run the output through a json mask
  - Added request-agent.js to verify there are no left-over connections that haven't closed. This is useful when making sure all connections have closed before shutting down a server.
  - Added a package.json file to define the module and its dependencies

chroma-externals: 15405/6

Change-Id: Ic243e7c48aa4816132c3699a863955d6cbd10bfa
Signed-off-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-on: http://review.whamcloud.com/15406
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>