HYD-4292 Create dirp
There are many cases in which we need to load multiple
dependencies within a directory. A lits of fixtures in a
test directory is one such example. Rather than require
each file within the directory individually, it makes
more sense to load the directory into an object and be
able to access each javascript or json file as a property
on the object. No matter how many files exist in the
folder, we only need to make one call to load them all.

The goal of dirp is to load all javascript and json
files within a specified directory into an object.
The directory name is passed in. All files in the
directory will be loaded except for the list of files
in the excluded list. Once the list of files is loaded,
it will simply require each file individually.

  - Added jasmine for testing
  - Added index file to load files in the given directory
  - Added package.json file to define module and
    dependencies
  - Added tests

Change-Id: I9e300b57c040e5a82b2323957df341c219a3f30c
Signed-off-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-on: http://review.whamcloud.com/15384
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>