HYD-4312 Upgrade testing fails due to __cmp__ of VersionInfo being misimplemented
VersionInfo which is used to compare package versions overrides _cmp_ however it
inherits from namedtuple which implements all of the rich comparisons like _lt_
and _gt_ and these rich comparisons take precedence over the _cmp_ method.

What is required is a version of VersionInfo that only includes _cmp_ and
includes some unit test to make sure it actually works!

This patch turns VersionInfo into a standalone class with tests. The tests are
not landed as part of this patch because of some CI issues, which means they will
follow in a later patch.

The move of the class into a dedicated file has a couple of advantages. Firstly it
can be used from multiple components and second it allows the pairing of code and tests
to be more obvious.

At present VersionInfo cannot be used from chroma_agent (lustre.py) because it is
not seriablizable as json. Fixing that seemed like feature creep.

Change-Id: I6d762cec4b723ec644730e7da927537d4436eb10
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/14267
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/14365
Reviewed-on: http://review.whamcloud.com/14652