HYD-4395 DNE Automated REST Interface Testing
Create a number of testing as part of the shared integration
test suite that do the following functionality via the rest interface.
The tested functionality should include:

  Create FS with single MDTs then add an MDT

  Create FS with multiple MDTs

  Add additional MDT’s

  Delete MDT’s (Commented out because of LU-6586)

  Re-add MDT’s

  Check that MDT0 cannot be deleted.

  Check that MDT1 cannot be deleted because of LU-6586.

  In all cases use lfs mkdir to check that the mdt can be used to

  create/delete files and check all stats in all cases.

Created a new file called test_filesystem_dne.py which allow for adding
and removing mdts. Although remove does not work in Lustre today the
functionality is correct in the tests and was the test code that discovered
the lustre limitation.  The remote operations class was expanded to
include the ability to write data and files to a specific mdt and written
in such a way the multiple mdts are exercised in a parallel. A problem with
very parallel ssh meant that mdts are parallelised but the file creation is
not. The delete mdt test is disabled with skip decorator.

  - Extend RemoteOperations to include exercise of the mdt.
  - Add test_filesystem_dne.py to include the new dne tests.
  - Add a catch so that requests may be used on a mac which does not have
    the latest requests module
  - The target endpoint was expanded to return the index of the target.

Change-Id: I4dd27931c43bcc953c18b68e5fdf95317c1660ab
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/14728
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>
Reviewed-by: Paul Kline <paul.everettx.kline@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
(cherry picked from commit 846d6599f02115d6fd54f3d833a5f33ca79a130c)
Reviewed-on: http://review.whamcloud.com/15101