HYD-4368 Support DNE Phase I in the IML GUI and Backend.
This means:

1) Be able to select N MDTs in the create filesystem screen.
2) Be able to add more MDTs to an existing FS.
3) Be able to remove non 0 MDTs.

The changes are minimal as IML was pretty much already written to
handle multiple MDT's. The code required mostly removal of exceptions
and a little bit of generalisation.

The unit tests have been expanded to do some multiple MDT testing.
Selenium testing as well was expanded to test the multiple MDT case.
This change requires the simultaneous landing of two tickets HYD-4368
and HYD-4320.

HYD-4368 is the backend changes HYD-4320 is the frontend changes

Change-Id: I719e679bb9b173d136a67d5038d05c645181cb8e
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/14327
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-on: http://review.whamcloud.com/14654
Tested-by: Jenkins
Reviewed-on: http://review.whamcloud.com/14445