HYD-4320 Support DNE Phase I in the IML GUI and Backend.
This means: 

1) Be able to select N MDTs in the create filesystem screen.
2) Be able to add more MDTs to an existing FS. 
3) Be able to remove non 0 MDTs. The changes are minimal as IML was
pretty much already written to handle multiple MDT's.

The code required mostly removal of exceptions and a little bit of
generalisation. The unit tests have been expanded to do some multiple
MDT testing.

Selenium testing as well was expanded to test the multiple MDT case.
This change requires the simultaneous landing of two tickets HYD-4368
and HYD-4320.

HYD-4368 is the backend changes HYD-4320 is the frontend changes. 

HYD-850 is also fixed as part of the front-end changes.

Change-Id: I834b06050388942b23336f589bc5e96e3045298d
Reviewed-on: http://review.whamcloud.com/14326
Tested-by: Jenkins
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Paul Kline <paul.everettx.kline@intel.com>
Tested-by: Paul Kline <paul.everettx.kline@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/14653