HYD-4321 Update Requests module due to security update
Requests 2.6.0, released March 14 fixes the security vulnerability
detailed in CVE-2015-2296.

Need to disable urllib3 warnings in tests because we use a self-
signed certificate but don't give the tests the certificate to
use in verifying the connection.   We should provide the certificate
as part of test but that's outside of the scope of this update and
should be handled in a different patch.

Signed-off-by: Joshua Kugler <joshua.kugler@intel.com>
Change-Id: Idc7930fcde1ca63e19fdf729ec59cb8f93e37079
Reviewed-on: http://review.whamcloud.com/14379
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>