HYD-4255 Turn corosync into a stateful object
We have a number of problems with corosync and pacemaker
where some additional knowledge, feedback and control
options would make the setup more reliable and usable
for the user.

Examples are: reconfiguration, failure after configuration,
failure to configure, use options etc etc.

The correct IML solution for this is to make them stateful
objects that are controlled using the statemachine.

This object would then have a relationship with other objects
like hosts, which would require corosync and pacemaker to
be active before the filesystem could be bought up.

This patch is an example of how not a land a large change. It
has grown into a number of interdependent changes that means
it has taken much to long to land.

The changes are numerous.

 - Creation of stateful objects for
     Corosync
     Pacemaker
     NTPConfiguration
     RSyslogConfiguration
 - Addition of tests for corosync
 - Addition of tests for pacemaker
 - Provision, although not enabled, of ability to detect corosync
   hosts that have the incorrect number of peers.
 - Separation of the models into more files to provide granularity.
 - Rationalization and separation of test utilities to provide better
   usability.

Change-Id: I62eed2feb74565aaff3b8332c013f23c8216da7e
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/17200
Tested-by: Jenkins
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>