HYD-4412 Updated Python requests module causes test aggregation to fail
An update python requests module, required to fix a security update 
CVE-2015-2296 causes errors in the test aggregation script.

The solution is to turn off the warnings from the module. It is probably
the case that we should also remove the source of the warnings and then
turn the warnings back on.

Change-Id: I4e64fdc0d446dfaf557798b92e021131db953908
Reviewed-on: http://review.whamcloud.com/14767
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>