HYD-4431 Create FP module
We are using lots of functional concepts when building the GUI.

Unfortunately, tools such as lodash fall short of ideal when trying to
do fp, (no autocurry, collection first). To shim this we have created
lodash-mixins, a module that relies on lodash and gives us those
properties for frequently used operations. The issue with this
approach is that we need to shim in a lib or flip at runtime, and on
top of that we are relying on an external dependency.

There are tools such as ramda that come closer to the functional
ideal. However, they are a grab-bag of functionality and an extra
external dep; We will still be limited by a lib and whatever hidden
edges it contains. Additionally, we should be able to build complex
behaviors out a few well designed transforms, i.e. a small core will
make for better and clearer composition of parts and reduce the need
to scan a page of a hundred methods to pick from.

Given all of this, create a module that does base fp operations that
has no external dependencies. This module will form the core of
higher-level operations going forward.

chroma-externals: 14845/4

Change-Id: I67070535204e7c49ca04c6d03fb24497f597ba61
Reviewed-on: http://review.whamcloud.com/14846
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>