HYD-4432 Update to Lustre build version 2.5.37
Update to the Lustre version 2.5.34.1 by changing
the following build parameters.

    JENKINS_JOB_NAME       = lustre-b_ieel2_0
    JENKINS_BUILD_NUMBER   = 185
    LUSTRE_RELEASE         = 2.5.37
    EL_KERNEL_RELEASE      = 2.6.32-504.16.2.el6
    SL11_KERNEL_RELEASE    = 3.0.101-0.47.52

The following chroma-dependencies are effected

    lustre-server-{el6,sles11}
    lustre-client-{el6,sles11}
    zfs
    xeon-phi-client

Refactored the Lustre dependencies addition process to interrogate
all of the various verison values from Jenkins given just a Jenkins
build name and number to make adding more O/S versions easier.

Cleaned up the use of a flawed proxy (lvfs -- which went away in
2.5.35) for unloading all of the modules that the lnet module
depends on and now properly remove the actual dependencies of the
lnet module, not some stand in for them.

Also renamed the various lustre server and client
chroma-dependencies and chroma-bundles in a consistent manner.
It might even be useful to start putting them into a directory
hierarchy.

chroma-externals: 14877/1
Change-Id: Ic5973960eef5b804f50b275b66eefad73801679c
Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-on: http://review.whamcloud.com/14882
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>