HYD-4434 Updated occurrence of RHEL 6.5 in test config files
A previous ticket HYD-4389 (63f9ebfdad219cd6aa7d371fec681829fed535b3)
change the configuration files to use TEST_DISTRIBUTION as a placeholder
for the required distribution however this change missed a few of the
configuration files namely selenium_cluster_config.json,
services_cluster_config.json and simulator.json. This need to be updated
to make use of the TEST_DISTRIBUTION placeholder along with required
changes to the code do do the substitution.

The substition of the files was straight-foward and the code to make
the dynamic update was present for all but simulator.json and so this was
added to the simulator jenkins_steps. There other other files in the
translator_tests directory which contain hardcoded distro versions but
these should not be updated as they are tests for the infrastructure and
the values are effectively mocks.

  - Modified the json files to contain TEST_DISTRIBUTION
  - Update the steps for the simulator to do the substition.

Change-Id: I5761df27b5ff03d1d74f51f66fc29c1788935c91
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/14913
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>