HYD-4437 Use provisioner-ng for EFS tests
The provisioner-ng hardware pool is larger and easier to manage,
allowing us to grow it more easily so we want to allow EFS take
advantage of that.

This patch also fixes a potential false positive failure in the many
iterations of this test across the DNE/HA/backing-store type loops.

Plus a fix an in the unmount code path where address was being used
rather than nodename. This bug was highlighted by the switch to
provisioner-ng.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: I4349567032bae53e6528ef8c0fffbe1c579b5453
Reviewed-on: http://review.whamcloud.com/14822
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Mathieu Drapeau <mathieu.drapeau@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/15013
Tested-by: Jenkins