HYD-4477 Add more methods to fp
This patch adds three useful methods, safe, eqFn, and noop to the
funtional programming module.

safe provides a mechanism to wrap function calls and return a default
value if any incoming arguments are nullish or the function body
throws an error. This is particularlly helpful in composition
and we can continue on with a default for expected error conditions.

eqFn takes two functions fnA, fnB, and applies them to two values,
a and b. It then uses strict equality comparison to determine if they
are the same value. This is useful to compare a nested value in two
objects with different interfaces.

noop is a "no operation" function. It is useful in cases where
you may want to exectute some code conditionally. It's also useful
for stubbing out external behaviors in test.

Change-Id: I1326913f31a30d7cc6be9bfdeb93534f7e24a352
Reviewed-on: http://review.whamcloud.com/15103
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>