HYD-4478 chroma-externals sometimes is not set to a pristine state
Is a checkout or update on a git repo is interupted the repo can end
up in a bit of a mess. Files will be consider modified/deleted/created
and further updates will fail or leave the files modified meaning that
the built product will be incorrect or fail.

The solution is to before every build do a

git reset --hard
git clean -dfx

The first will remove any known changes from the repo and the second
will remove any unknown changes from the repo. The means that we are
guaranteed to be started from a known position.

To ensure that this reset to a known state at the beginning of a build
does not hide bad behaviour of previous builds the chroma-externals an
additional validation that chroma-externals is unmodified at end of a
build has been added.

Change-Id: I75eefb1a539598c8fdbbdd35ab310ea8c2bf18da
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/15107
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>