HYD-4502 Switch on description in step titles
step.description will soon contain useful information to a user as the step proceeds. This ticket checks if step description contains a sentence useful to a human and if so displays that.

We check step.description to see if it does not start with step.class_name in both the old and new GUIs. If so, we use the description

- Switch on step.description in old and new GUIs

Signed-off-by: Joe Grund <joe.grund@intel.com>
Change-Id: I218fe78f423d69d110b10078a579f63008207e47
Reviewed-on: http://review.whamcloud.com/15212
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>