HYD-4394 DNE Automated CLI Interface Testing
Create a number of testing as part of the shared integratiion
test suite that do the following functionality via the rest interface.

This test doesn't contain the delete MDT because the is not possible yet.

Create FS with single MDTs then add an MDT
Check that MDT0 cannot be deleted.
Check that MDT1 cannot be deleted because of LU-6586.

This does not test mdt functionality as that is covered in the rest
interface testing.

Change-Id: I2a9dfe592669ab5bcbe278f2215d2aec14326a94
Reviewed-on: http://review.whamcloud.com/15249
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>
(cherry picked from commit f65e5bf49157d2a88e653a20a721277b29d08a7a)
Reviewed-on: http://review.whamcloud.com/15326