HYD-4511 get_label error when back from failover mdt
If a filesystem is redetected then some of the existing targets
can disappear because at the time of redetection they, or to be
more precise the primary target, are not present for some reason
This causes IML to delete the target record for those missing
but once present targets leading to a number of errors including
a get_label() error.

The solution to this problem is to only consider new (not previously
known targets for deletion by analysing the 'new targets list' rather
than all targets for the deletion if primary not found question.

  - Renamed created_mgts to created_mgss because it is more consistant
    with the rest of the code.
  - Change scan to look at all created targets rather than all targets
    which actually results in cleaner more consistant code

Change-Id: I939851fab0c7fe615668072c7338a4a224f6fc29
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/15268
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>