HYD-4515 Add an invoke function to the fp library
Enhance the fp library by adding an invoke function. This will allow us
to invoke a function in a points free manor when composing functions
together.

The solution was to add an invoke function to the fp library. This is a
curried function that receives a function and a list of args. The
function is then invoked with the list of arguments passed.

  - Added fp to src globs after all angular modules are loaded
  - Added the invoke function to the fp library
  - Added jasmine-n-matchers to check how many times a function is
    called in a test
  - Added tests to test the invoke function


Change-Id: I8c67239ebeaff710a3dd282a655e02a4b61e1416
Signed-off-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-on: http://review.whamcloud.com/15283
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Tested-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>