HYD-4523 Enforce upgrade of python-requests
When we updated python-requests to 2.6.0 in IML 2.2.0.0 we
did not enforce the upgrade on existing systems by requiring
python-requests >= 2.6.0. The spec files for the rpm builds
need to be updated to enforce this requirement.

As per the description update the spec files appropriately.

  - Updated chroma-agent/chroma-agent.spec
  - Updated chroma-manager/chroma-manager.spec
  - Updated cluster-sim/chroma-cluster-sim.spec

Change-Id: I6f113e242925b514c4b362491dcef4875f609bcc
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/15337
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
(cherry picked from commit a326770aafd8dc4c128d3f95e516231ebd54aa2d)
Reviewed-on: http://review.whamcloud.com/15376