HYD-4501 System does not recover when an ethernet interface is restored
Bringing the eth0 interface down with "ifdown eth0" results in the expected
result of an alert that contact was lost with lotus-25.iml.intel.com
and lotus-25 is not stonithed. Bringing eth0 back up by ssh'ing into the
HA paired server and then into eth1 and running "ifup eth0" brings the
interface back up as expected. I am able to ssh into lotus-25 on eth0 again.
The GUI does not detect that the interface is active and continues to show
the lotus-25 contact lost alert.

The problem here is that the CryptoClient class only captures certain exceptions
returned by the requests package and unknown exceptions cause the thread to
exit and never restart.  The agent runs 2 message threads one pushing messages
to the manager (HttpWriter) and on pulling them from the manager (HttpReader)
if either of these exits then the agent and the manager have lost communication.
The solution is to add the missing exception types to the list of caught exceptions
and change the behaviour of the unexcepted exception to be different depending
on the context.

If the agent is running in debug (which in particular means CI testing) then
the uncaught exception should be raised again causing the thread to exit and the
application to fail.  If the agent is running in a production mode (!debug) then
information should be logged but the thread should continue to run so that the
application continues correctly for the user. Bad data will not be passed and
the user may see an error, but it should be recoverable by a user.

  - Catch additional ReadTimeout and SSLError exceptions
  - Catch unexpected exceptions and deal with in a context sensitive way
  - Create a new flag logging_in_debug_mode to allow code to know if debug is
    enabled.

Change-Id: Ic2e04093703293affeee639ada4031e4996da0d3
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/15350
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>